colocate `useLatestAnswerForPatientQuery`
