--- packages/web/app/api/queries/useLatestAnswerForPatientQuery.js
@@ -1,22 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { useApi } from '../useApi';
-import { isErrorUnknownAllow404s } from '../TamanuApi';
-
-export const useLatestAnswerForPatientQuery = (patientId, dataElementCode) => {
-  const api = useApi();
-
-  return useQuery(
-    ['survey', patientId, dataElementCode],
-    () =>
-      api.get(
-        `surveyResponseAnswer/latest-answer/${encodeURIComponent(dataElementCode)}`,
-        {
-          patientId,
-        },
-        { isErrorUnknown: isErrorUnknownAllow404s },
-      ),
-    {
-      enabled: !!patientId && !!dataElementCode,
-    },
-  );
-};
